test: split the ingest driver tests by program
The module covering both command lines and the worker between them went
over the thousand-line convention when this round's tests landed in it.
The worker's own tests move to a module of their own, and what runs
either program -- neither half's tests can do without the other's -- to a
helpers module beside them, so that neither test module imports the
other.
